add client config
