+++ rocketmq-go/model/query_result.go
