+++ rocketmq-go/model/topic_publishInfo.go
